services/noesis-bridge/requirements.txt
PyYAML==6.0.2
openai
